+++ DEBtoolPyIF/estimation/runner.py
@@ -1,67 +1,12 @@
-import functools
 import os
-import matlab.engine
+from .wrapper import DEBtoolWrapper
 from io import StringIO
 
 
-def run_estimation(run_files_dir, species_name, window=False):
-    eng = matlab.engine.start_matlab()
-    if window:
-        eng.desktop(nargout=0)
-
-    # Check that file exists
-    if not os.path.isfile(f"{run_files_dir}/run_{species_name}.m"):
-        raise Exception("run.m file does not exist.")
-
-    # Change MATLAB working directory
-    eng.cd(run_files_dir, nargout=0)
-
-    # Run estimation
-    eng.eval(f'run_{species_name}', nargout=0)
-
-    # Load results .mat file and return parameters and estimation errors.
-    eng.eval(f"load('{run_files_dir}/results_{species_name}.mat')", nargout=0)
-    pars = eng.workspace['par']
-    estimation_errors = eng.workspace['metaPar']
-    meta_data = eng.workspace['metaData']
-    # Close engine
-    eng.quit()
-    return pars, estimation_errors, meta_data
-
-
-class EstimationRunner:
+class EstimationRunner(DEBtoolWrapper):
     # TODO: Store empty buffers for output as class attributes so that hide_output can be a method option
 
-    def __init__(self, matlab_session=None, window=False, clear_before=True, species_name=None):
-        if matlab_session is None:
-            self.eng = matlab.engine.start_matlab()
-        if matlab_session == 'find':
-            matlab_sessions = matlab.engine.find_matlab()
-            if not len(matlab_sessions):
-                raise Exception("No shared MATLAB session found. Make sure to run 'matlab.engine.shareEngine' in the "
-                                "MATLAB command window.")
-            else:
-                self.eng = matlab.engine.connect_matlab(matlab_sessions[0])
-        else:
-            self.eng = matlab.engine.connect_matlab(matlab_session)
-        self.window = window
-        self.clear_before = clear_before
-        self.species_name = species_name
-
-    def apply_options_decorator(func):
-        @functools.wraps(func)
-        def wrapper(self, *args, **kwargs):
-            # Apply the options just like in the original apply_options method
-            if self.window:
-                self.eng.desktop(nargout=0)
-            if self.clear_before:
-                self.clear()
-            # Call the actual function
-            return func(self, *args, **kwargs)
-
-        return wrapper
-
-    @apply_options_decorator
+    @DEBtoolWrapper.apply_options_decorator
     def run_estimation(self, run_files_dir: str, hide_output=True):
         """
 
@@ -90,7 +35,7 @@ class EstimationRunner:
 
         return success
 
-    @apply_options_decorator
+    @DEBtoolWrapper.apply_options_decorator
     def fetch_pars_from_mat_file(self, run_files_dir: str, results_file=None):
 
         if results_file is None:
@@ -105,7 +50,7 @@ class EstimationRunner:
 
         return pars
 
-    @apply_options_decorator
+    @DEBtoolWrapper.apply_options_decorator
     def fetch_errors_from_mat_file(self, run_files_dir: str, error_type='RE', results_file=None):
         if error_type not in ('RE', 'SAE', 'SSE'):
             raise Exception(f"Invalid error_type {error_type}. Error type must be either 'RE', 'SAE' or 'SSE'.")
@@ -123,7 +68,7 @@ class EstimationRunner:
 
         return errors
 
-    @apply_options_decorator
+    @DEBtoolWrapper.apply_options_decorator
     def fetch_data_from_mat_file(self, run_files_dir: str, results_file=None):
         if results_file is None:
             results_file = f"results_{self.species_name}.mat"
@@ -135,7 +80,7 @@ class EstimationRunner:
 
         return self.eng.workspace['data']
 
-    @apply_options_decorator
+    @DEBtoolWrapper.apply_options_decorator
     def fetch_predictions_from_mat_file(self, run_files_dir: str, results_file=None):
         if results_file is None:
             results_file = f"results_{self.species_name}.mat"
@@ -147,14 +92,6 @@ class EstimationRunner:
 
         return self.eng.workspace['prdData']
 
-    def cd(self, workspace_dir):
-        self.eng.cd(os.path.abspath(workspace_dir), nargout=0)
-
-    def close(self):
-        self.eng.quit()
-
-    def clear(self):
-        self.eng.eval("clear", nargout=0)
 
 if __name__ == '__main__':
     print('Done!')
